added alternative with wakelock and execution in onCreate directly
